+++ src/Data/Bytes.hs
@@ -23,7 +23,7 @@ module Data.Bytes
   , Pure.emptyPinned
   , Pure.emptyPinnedU
     -- * Properties
-  , null
+  , Pure.null
   , Pure.length
     -- * Decompose
   , uncons
@@ -47,6 +47,9 @@ module Data.Bytes
   , dropWhile
   , takeWhileEnd
   , dropWhileEnd
+    -- * Traversals
+  , Pure.map
+  , Pure.mapU
     -- * Folds
   , Pure.foldl
   , Pure.foldl'
@@ -79,6 +82,9 @@ module Data.Bytes
   , intercalateByte2
   , concatArray
   , concatArrayU
+    -- * Searching
+  , replace
+  , findIndices
     -- * Counting
   , Byte.count
     -- * Prefix and Suffix
@@ -116,9 +122,10 @@ module Data.Bytes
   , Pure.fnv1a32
   , Pure.fnv1a64
     -- * Unsafe Slicing
-  , unsafeTake
+  , Pure.unsafeTake
   , Pure.unsafeDrop
-  , unsafeIndex
+  , Pure.unsafeIndex
+  , Pure.unsafeHead
     -- * Copying
   , Pure.unsafeCopy
     -- * Pointers
@@ -153,24 +160,24 @@ module Data.Bytes
   , unlift
   ) where
 
-import Prelude hiding (length,takeWhile,dropWhile,null,foldl,foldr,elem,replicate,any,all,readFile)
+import Prelude hiding (length,takeWhile,dropWhile,null,foldl,foldr,elem,replicate,any,all,readFile,map)
 
 import Control.Monad.Primitive (PrimMonad,primitive_,unsafeIOToPrim)
 import Control.Monad.ST.Run (runByteArrayST)
 import Cstrlen (cstringLength#)
-import Data.Bits((.&.),(.|.),shiftL,finiteBitSize)
-import Data.Bytes.Pure (length,fromByteArray,foldr,unsafeDrop)
-import Data.Bytes.Types (Bytes(Bytes,array,offset))
 import Data.ByteString.Short.Internal (ShortByteString(SBS))
-import Data.Maybe (fromMaybe)
+import Data.Bytes.Pure (length,fromByteArray,foldr,unsafeDrop)
+import Data.Bytes.Pure (unsafeIndex)
+import Data.Bytes.Types (Bytes(Bytes,array,offset))
 import Data.Primitive (Array,ByteArray(ByteArray))
 import Data.Text.Short (ShortText)
 import Foreign.C.String (CString)
 import Foreign.Ptr (Ptr,plusPtr,castPtr)
 import GHC.Exts (Addr#,Word#,Int#)
 import GHC.Exts (Int(I#),Ptr(Ptr))
-import GHC.Word (Word8(W8#),Word32)
+import GHC.Word (Word8(W8#))
 import Reps (Bytes#(..),word8ToWord#)
+import Data.Bytes.Search (findIndices,replace,isInfixOf)
 
 import qualified Data.Bytes.Byte as Byte
 import qualified Data.Bytes.Chunks as Chunks
@@ -186,11 +193,6 @@ import qualified Data.Primitive as PM
 import qualified Data.Primitive.Ptr as PM
 import qualified Data.Text.Short as TS
 import qualified GHC.Exts as Exts
-
--- | Is the byte sequence empty?
-null :: Bytes -> Bool
-{-# inline null #-}
-null (Bytes _ _ len) = len == 0
 
 -- | Extract the head and tail of the 'Bytes', returning 'Nothing' if
 -- it is empty.
@@ -245,67 +247,6 @@ isSuffixOf (Bytes a aOff aLen) (Bytes b bOff bLen) =
     then compareByteArrays a aOff b (bOff + bLen - aLen) aLen == EQ
     else False
 
--- | Is the first argument an infix of the second argument?
--- 
--- Uses the Rabin-Karp algorithm: expected time @O(n+m)@, worst-case @O(nm)@.
-isInfixOf :: Bytes -- ^ String to search for
-          -> Bytes -- ^ String to search in
-          -> Bool
-isInfixOf p s = null p || (not . null) (snd $ breakSubstring p s)
-
-breakSubstring :: Bytes -- ^ String to search for
-               -> Bytes -- ^ String to search in
-               -> (Bytes,Bytes) -- ^ Head and tail of string broken at substring
-breakSubstring pat =
-  case lp of
-    0 -> (mempty,)
-    1 -> breakByte (unsafeHead pat)
-    _ -> if lp * 8 <= finiteBitSize (0 :: Word)
-             then shift
-             else karpRabin
-  where
-  unsafeSplitAt i s = (unsafeTake i s, unsafeDrop i s)
-  lp                = length pat
-  {-# INLINE breakByte #-}
-  breakByte b bytes = fromMaybe (mempty,bytes) $ Byte.split1 b bytes
-  {-# INLINE karpRabin #-}
-  karpRabin :: Bytes -> (Bytes, Bytes)
-  karpRabin src
-      | length src < lp = (src,mempty)
-      | otherwise = search (rollingHash $ unsafeTake lp src) lp
-    where
-    k           = 2891336453 :: Word32
-    rollingHash = Pure.foldl' (\h b -> h * k + fromIntegral b) 0
-    hp          = rollingHash pat
-    m           = k ^ lp
-    get = fromIntegral . unsafeIndex src
-    search !hs !i
-        | hp == hs && pat == unsafeTake lp b = u
-        | length src <= i                    = (src,mempty) -- not found
-        | otherwise                          = search hs' (i + 1)
-      where
-      u@(_, b) = unsafeSplitAt (i - lp) src
-      hs' = hs * k +
-            get i -
-            m * get (i - lp)
-  {-# INLINE shift #-}
-  shift :: Bytes -> (Bytes, Bytes)
-  shift !src
-      | length src < lp = (src,mempty)
-      | otherwise       = search (intoWord $ unsafeTake lp src) lp
-    where
-    intoWord :: Bytes -> Word
-    intoWord = Pure.foldl' (\w b -> (w `shiftL` 8) .|. fromIntegral b) 0
-    wp   = intoWord pat
-    mask = (1 `shiftL` (8 * lp)) - 1
-    search !w !i
-        | w == wp         = unsafeSplitAt (i - lp) src
-        | length src <= i = (src, mempty)
-        | otherwise       = search w' (i + 1)
-      where
-      b  = fromIntegral (unsafeIndex src i)
-      w' = mask .&. ((w `shiftL` 8) .|. b)
-
 -- | Find the longest string which is a prefix of both arguments.
 longestCommonPrefix :: Bytes -> Bytes -> Bytes
 longestCommonPrefix a b = loop 0
@@ -315,7 +256,7 @@ longestCommonPrefix a b = loop 0
     | into < maxLen
       && unsafeIndex a into == unsafeIndex b into
       = loop (into + 1)
-    | otherwise = unsafeTake into a
+    | otherwise = Pure.unsafeTake into a
   maxLen = min (length a) (length b)
 
 -- | Create a byte sequence with one byte.
@@ -418,30 +359,19 @@ elemLoop !r !w (Bytes arr@(ByteArray arr# ) off@(I# off# ) len) = case len of
 -- | Take bytes while the predicate is true.
 takeWhile :: (Word8 -> Bool) -> Bytes -> Bytes
 {-# inline takeWhile #-}
-takeWhile k b = unsafeTake (countWhile k b) b
+takeWhile k b = Pure.unsafeTake (countWhile k b) b
 
 -- | Drop bytes while the predicate is true.
 dropWhile :: (Word8 -> Bool) -> Bytes -> Bytes
 {-# inline dropWhile #-}
-dropWhile k b = unsafeDrop (countWhile k b) b
-
--- | Index into the byte sequence at the given position. This index
--- must be less than the length.
-unsafeIndex :: Bytes -> Int -> Word8
-{-# inline unsafeIndex #-}
-unsafeIndex (Bytes arr off _) ix = PM.indexByteArray arr (off + ix)
-
--- | Access the first byte. The given 'Bytes' must be non-empty.
-{-# inline unsafeHead #-}
-unsafeHead :: Bytes -> Word8
-unsafeHead bs = unsafeIndex bs 0
+dropWhile k b = Pure.unsafeDrop (countWhile k b) b
 
 -- | /O(n)/ 'dropWhileEnd' @p@ @b@ returns the prefix remaining after
 -- dropping characters that satisfy the predicate @p@ from the end of
 -- @t@.
 dropWhileEnd :: (Word8 -> Bool) -> Bytes -> Bytes
 {-# inline dropWhileEnd #-}
-dropWhileEnd k !b = unsafeTake (length b - countWhileEnd k b) b
+dropWhileEnd k !b = Pure.unsafeTake (length b - countWhileEnd k b) b
 
 -- | /O(n)/ 'takeWhileEnd' @p@ @b@ returns the longest suffix of
 -- elements that satisfy predicate @p@.
@@ -450,12 +380,6 @@ takeWhileEnd :: (Word8 -> Bool) -> Bytes -> Bytes
 takeWhileEnd k !b =
   let n = countWhileEnd k b
    in Bytes (array b) (offset b + length b - n) n
-
--- | Take the first @n@ bytes from the argument. Precondition: @n ≤ len@
-unsafeTake :: Int -> Bytes -> Bytes
-{-# inline unsafeTake #-}
-unsafeTake n (Bytes arr off _) =
-  Bytes arr off n
 
 -- Internal. The returns the number of bytes that match the
 -- predicate until the first non-match occurs. If all bytes